requirements.txt
mcp[cli]>=1.0.0
caldav>=1.3.0
icalendar>=5.0.0
sqlalchemy>=2.0.0
uvicorn>=0.20.0
starlette>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.0.0
httpx>=0.27.0
pytz>=2023.3
python-dateutil>=2.8.2
